prologue: forcefully reload sfc dependency
sfc_driverlink may be left loaded in a given system for a long time,
while the source-built sfc slowly goes out of sync with it. At some
point, a new version of sfc may refuse to load, breaking the testing.
Fix this by forcefully reloading sfc_driverlink along with sfc.

Signed-off-by: Viacheslav Galaktionov <viacheslav.galaktionov@arknetworks.am>