usecases/runtime_tx_queue_setup: classify i40e failures
In the case of net/i40e, going for runtime Tx queue setup
causes failures on device start. Provide TRC expectations.

When the bug has been fixed, the test will fail in a
different way because of a bug in deferred Tx queue
start. Assuming the partial fix has been applied
for that bug too, provide updates for the test
which help to classify remaining regressions.

Signed-off-by: Ivan Malov <ivan.malov@oktetlabs.ru>
Acked-by: Viacheslav Galaktionov <viacheslav.galaktionov@arknetworks.am>
Acked-by: Denis Pryazhennikov <denis.pryazhennikov@arknetworks.am>