usecases/rx_intr: poll for inbound traffic in batches
Test results for i40e suggest that the driver enables
vectorised Rx by default. In that mode, Rx burst size
must be equal to or greater than a certain batch size.
Otherwise, Rx burst comes up empty even when a packet
has hit the NIC. Revisit the test to comply with that.

Signed-off-by: Ivan Malov <ivan.malov@oktetlabs.ru>
Reviewed-by: Andrew Rybchenko <andrew.rybchenko@oktetlabs.ru>