xmit/one_packet: improve outer UDP checksum handling for TSO
A PMD may support tunnel TSO but lack support for outer
UDP checksum offload. If so, when outer frame uses IPv4,
the application should set the UDP checksum to 0 in TSO
superframe. Resulting packets should preserve the value.

This amendment improves the test and will help to avoid
regressions in i40e runs when the PMD has been fixed to
quit advertising outer UDP offload it cannot do in fact.

Signed-off-by: Ivan Malov <ivan.malov@oktetlabs.ru>
Reviewed-by: Andrew Rybchenko <andrew.rybchenko@oktetlabs.ru>