usecases/xstats_dev_state: fix comparison on number of names
The comment on the DPDK operation `rte_eth_xstats_get_names` states
that this operation returns the number of entries that should be given
for it if no enough entries to store names has been provided.

Despite it the test compares this number against the number of names
actually filled if enough number of entries has been provided.

This comparison is wrong for the DPDK Virtio driver. For example,
if the test tests the `CONFIGURED` state of this device it fails with
the following error:
```
Bad number of all xstat names obtained: 13; must be 38
```

Thus, update test to respect the comment on this DPDK operation.

OL-Redmine-Id: 12717
Signed-off-by: Boleslav Stankevich <boleslav.stankevich@oktetlabs.ru>
Reviewed-by: Andrew Rybchenko <andrew.rybchenko@oktetlabs.ru>