usecases/tx_desc_nb: preserve host fdb on virtio-virtio
The virtio-virtio configuration uses a Linux bridge to connect TA
guests. This bridge uses a forwarding database to learn addresses
and to use them to reduce unicast flooding.
Some time is required to let the bridge to forget an address learnt
previously.

Thus, use addresses consistent with addresses for other tests.

OL-Redmine-Id: 12807
Signed-off-by: Boleslav Stankevich <boleslav.stankevich@oktetlabs.ru>
Reviewed-by: Andrew Rybchenko <andrew.rybchenko@oktetlabs.ru>