usecases/flow_ctrl_set: expect right autonegotiation outcome
The test does not account for the logic IEEE 802.3 envisages for different
Link Code Word combinations (for DUT and the link partner) to select final
flow control mode. Augment the test with some minimal logic to fix results
on AMD Solarflare adaptors. Also, add proper initialisation of the partner.

Leave it up to the original author to add extra parameters for some custom
initialisation of the link partner and adjust the expectation logic.

Fixes: 09a665cb14d8 ("usecases: add Flow Control set test")
Signed-off-by: Ivan Malov <ivan.malov@arknetworks.am>
Acked-by: Viacheslav Galaktionov <viacheslav.galaktionov@arknetworks.am>