usecases/tx_desc_nb: cope with vFIFO drops on TxQ depth 2048
On the X4 <--> X2 rig, this iteration causes intermittent drops because of
the vFIFO at the Tester side not being able to cope with the faster sender,
which is reflected in the 'port_rx_pm_discard_vfifo_full' counter. To make
the iteration pass, increase the payload size within reasonable boundaries.

Signed-off-by: Ivan Malov <ivan.malov@arknetworks.am>
Reviewed-by: Viacheslav Galaktionov <viacheslav.galaktionov@arknetworks.am>