doc: support manual overrides for Sphinx builds
This allows to replace some of the documentation with either
translations or manual fixes.

This one is a bit of a hack, but it works for now. The key problem is
that index.rst is completely replaced (in case we want to remove/add
things) so if one adds a new file there in the tree it will be missing
from the resulting documentation.

Signed-off-by: Andrew Rybchenko <andrew.rybchenko@oktetlabs.ru>
Reviewed-by: Artem Andreev <artem.andreev@oktetlabs.ru>
Reviewed-by: Viacheslav Galaktionov <viacheslav.galaktionov@arknetworks.am>