prologue: fix failure because of net driver load failure
Network driver is not strictly required on IUT for DPDK testing.
The network driver could be specified in site-specific configuration
of test rigs, but not always available. Try to continue.

If AF_XDP is tested and network driver is not loaded, it will fail
anyway on attempt to bind tested PCI function to the network driver.

Signed-off-by: Andrew Rybchenko <andrew.rybchenko@oktetlabs.ru>
Reviewed-by: Daniil Byshenko <daniil.byshenko@oktetlabs.ru>
Reviewed-by: Viacheslav Galaktionov <viacheslav.galaktionov@arknetworks.am>